Fix 500 on every dashboard load: CLOSED_RANGE_REUSE_WINDOW_MS overflowed int4
findReusableSync casts its windowMs parameter with $5::int. That was
fine for the original 5-minute/1-hour windows, but
CLOSED_RANGE_REUSE_WINDOW_MS (100 years in milliseconds =
3,153,600,000,000) is far past Postgres int4's ~2.147 billion ceiling,
so every call for a closed date range - which, since the change that
introduced it, is effectively every dashboard load with a past end
date - failed with 22003 numeric_value_out_of_range
(pg_strtoint32_safe) and surfaced to the seller as a bare "Internal
server error". Confirmed live in the API log the user provided.

Changed the cast to $5::bigint, which comfortably holds the value (up
to ~9.2 * 10^18). The unit test suite mocks Postgres and never
actually type-checks a bound parameter against a real column/cast, so
this shipped without being caught - a gap worth remembering, not
fixing here.

Verified: node --check on server.js and the full test suite (154/154).
This one couldn't be verified against a real Postgres cast from this
sandbox either, so I'm relying on the read: int4 range is well
documented and 3.15e12 is unambiguously over it by three orders of
magnitude.

+++ apps/api/src/server.js
@@ -362,7 +362,7 @@ async function findReusableSync(tenantId, reportType, range, windowMs = SYNC_REU
   const recent = await pool.query(
     `select id, completed_at, range_start, range_end from sync_jobs
      where tenant_id=$1 and report_type=$2 and status='completed'
-       and completed_at is not null and completed_at > now() - ($5::int * interval '1 millisecond')
+       and completed_at is not null and completed_at > now() - ($5::bigint * interval '1 millisecond')
        and range_start is not null and range_end is not null
        and range_start <= $3 and range_end >= $4
      order by completed_at desc limit 1`,
